// src/app/models/skill.interface.ts
/**
 * Represents a skill
 *
 * @export
 * @interface ISkill
 */
export interface ISkill {
  /**
   * Unique identifier of the skill
   *
   * @type {number}
   * @memberof ISkill
   */
  id: number;

  /**
   * Name of the skill
   *
   * @type {string}
   * @memberof ISkill
   */
  name: string;
}